+++ rt/lastmodifier/browser/document_byline.py
@@ -1,16 +1,18 @@
 # -*- coding: utf-8 -*-
 
+from AccessControl import getSecurityManager
 from DateTime import DateTime
+from Products.CMFCore.utils import getToolByName
+from plone.app.layout.viewlets.content import DocumentBylineViewlet as BaseDocumentBylineViewlet
+from plone.memoize.view import memoize
+from rt.lastmodifier.browser.changenote import ShowChangeNoteViewlet
+from rt.lastmodifier.permissions import DocumentByLineViewAuthor
+from rt.lastmodifier.permissions import DocumentByLineViewChangeNote
+from rt.lastmodifier.permissions import DocumentByLineViewLastModifier
+from rt.lastmodifier.permissions import DocumentByLineViewModifiedDate
+from rt.lastmodifier.permissions import DocumentByLineViewPublishedDate
 from zope.component import getMultiAdapter
 from zope.interface import Interface
-from Products.CMFCore.utils import getToolByName
-from plone.memoize.view import memoize
-from plone.app.layout.viewlets.content import DocumentBylineViewlet as BaseDocumentBylineViewlet
-from rt.lastmodifier.permissions import DocumentByLineViewAuthor, DocumentByLineViewLastModifier, \
-                                        DocumentByLineViewModifiedDate, DocumentByLineViewPublishedDate, \
-                                        DocumentByLineViewChangeNote
-from rt.lastmodifier.browser.changenote import ShowChangeNoteViewlet
-from AccessControl import getSecurityManager
 
 
 class DocumentBylineViewlet(BaseDocumentBylineViewlet, ShowChangeNoteViewlet):
@@ -26,7 +28,7 @@ class DocumentBylineViewlet(BaseDocumentBylineViewlet, ShowChangeNoteViewlet):
 
     @memoize
     def show(self):
-        if self.can_see_author or self.can_see_last_modifier or\
+        if self.can_see_author or self.can_see_last_modifier or \
            self.can_see_modified or self.can_see_published:
             return True
         return False
@@ -61,3 +63,19 @@ class DocumentBylineViewlet(BaseDocumentBylineViewlet, ShowChangeNoteViewlet):
     def modifiername(self):
         modifier = self.modifier()
         return modifier and modifier['fullname'] or self.last_modifier()
+
+    def modification_date(self):
+        return self.context.ModificationDate()
+
+
+class DocumentBylineFolderViewlet(DocumentBylineViewlet):
+    """When on folders, last modifier and last modification date must be taken from
+    last modified content inside the folder
+    """
+
+    def modification_date(self):
+        catalog = getToolByName(self.context, 'portal_catalog')
+        results = catalog(sort_on='modified', sort_order='reverse', sort_limit=1)
+        if results:
+            return results[0].modified
+        return super(DocumentBylineFolderViewlet, self).modification_date()